fixing problem where when searching messages for delete it would break because it had no messages left to search

+++ commands/moderation/deleteMessages.js
@@ -18,39 +18,41 @@ module.exports = {
     const channel = interaction.channel;
     let messages = await channel.messages.fetch({ limit: 100 });
     let lastMessage = messages.last();
-    let messagesToDelete = messages.filter(
-      (message) =>
-        message.content.toLowerCase().includes(keyword.toLowerCase()) &&
-        message.id !== lastMessage.id
-    );
     let minimumMessages = 50;
+    let countDeletedMessages = 0;
 
-    for (const message of messagesToDelete.values()) {
-      console.log(message.content);
-      minimumMessages--;
-    }
-
+    await interaction.reply({
+      content: `Deletando mensagens que contenham a palavra chave ${keyword}...`,
+    });
+    let i = 0;
     while (minimumMessages > 0) {
-      messages = await channel.messages.fetch({
-        limit: 100,
-        before: lastMessage.id,
-      });
-      lastMessage = messages.last();
-      messagesToDelete = messages.filter(
-        (message) =>
-          message.content.toLowerCase().includes(keyword.toLowerCase()) &&
-          message.id !== lastMessage.id
-      );
+      if (lastMessage && lastMessage.id) {
+        if (i >= 1) {
+          messages = await channel.messages.fetch({
+            limit: 100,
+            before: lastMessage.id,
+          });
+          lastMessage = messages.last();
+        }
+        let messagesToDelete = messages.filter(
+          (message) =>
+            message.content.toLowerCase().includes(keyword.toLowerCase()) &&
+            message.id !== lastMessage.id
+        );
 
-      for (const message of messagesToDelete.values()) {
-        await message.delete();
-        minimumMessages--;
+        for (const message of messagesToDelete.values()) {
+          await message.delete();
+          countDeletedMessages++;
+          minimumMessages--;
+        }
+        i++;
+      } else {
+        break;
       }
     }
 
-    await interaction.reply({
-      content: "Mensagens apagadas com sucesso!",
-      ephemeral: true,
+    await interaction.editReply({
+      content: `Foram deletadas ${countDeletedMessages} mensagens!`
     });
-  },
+  }
 };
